Added tests and moved all test out of the web package since they are truelly integration tests.
